# gen/http/authentication_limiter/server/encode_decode.go
// Code generated by goa v3.11.0, DO NOT EDIT.
//
// authenticationLimiter HTTP server encoders and decoders
//
// Command:
// $ goa gen admin-panel/design

package server

import (
	authenticationlimiter "admin-panel/gen/authentication_limiter"
	"context"
	"io"
	"net/http"

	goahttp "goa.design/goa/v3/http"
	goa "goa.design/goa/v3/pkg"
)

// EncodeGetAuthenticationLimiterResponse returns an encoder for responses
// returned by the authenticationLimiter getAuthenticationLimiter endpoint.
func EncodeGetAuthenticationLimiterResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, interface{}) error {
	return func(ctx context.Context, w http.ResponseWriter, v interface{}) error {
		res, _ := v.(*authenticationlimiter.GetAuthenticationLimiterResult)
		enc := encoder(ctx, w)
		body := NewGetAuthenticationLimiterResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// EncodeSetAuthenticationLimiterResponse returns an encoder for responses
// returned by the authenticationLimiter setAuthenticationLimiter endpoint.
func EncodeSetAuthenticationLimiterResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, interface{}) error {
	return func(ctx context.Context, w http.ResponseWriter, v interface{}) error {
		res, _ := v.(*authenticationlimiter.SetAuthenticationLimiterResult)
		enc := encoder(ctx, w)
		body := NewSetAuthenticationLimiterResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeSetAuthenticationLimiterRequest returns a decoder for requests sent to
// the authenticationLimiter setAuthenticationLimiter endpoint.
func DecodeSetAuthenticationLimiterRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (interface{}, error) {
	return func(r *http.Request) (interface{}, error) {
		var (
			body SetAuthenticationLimiterRequestBody
			err  error
		)
		err = decoder(r).Decode(&body)
		if err != nil {
			if err == io.EOF {
				return nil, goa.MissingPayloadError()
			}
			return nil, goa.DecodePayloadError(err.Error())
		}
		payload := NewSetAuthenticationLimiterPayload(&body)

		return payload, nil
	}
}

// EncodeDelAuthenticationLimiterResponse returns an encoder for responses
// returned by the authenticationLimiter delAuthenticationLimiter endpoint.
func EncodeDelAuthenticationLimiterResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, interface{}) error {
	return func(ctx context.Context, w http.ResponseWriter, v interface{}) error {
		res, _ := v.(*authenticationlimiter.DelAuthenticationLimiterResult)
		enc := encoder(ctx, w)
		body := NewDelAuthenticationLimiterResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}