# gen/http/openapi.yaml
swagger: "2.0"
info:
    title: adminApiService
    description: Service for multiplying numbers, a Goa teaser
    version: ""
host: 0.0.0.0:18006
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
...
    /lpnode/lpnode_admin_panel/authenticationLimiter/setAuthenticationLimiter:
        post:
            tags:
                - authenticationLimiter
            summary: setAuthenticationLimiter authenticationLimiter
            description: set limit information
            operationId: authenticationLimiter#setAuthenticationLimiter
            parameters:
                - name: SetAuthenticationLimiterRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/AuthenticationLimiterSetAuthenticationLimiterRequestBody'
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/AuthenticationLimiterSetAuthenticationLimiterResponseBody'
            schemes:
                - http
...
    /lpnode/lpnode_admin_panel/relayAccount/delete:
        post:
            tags:
                - relayAccount
            summary: deleteAccount relayAccount
            operationId: relayAccount#deleteAccount
            parameters:
                - name: DeleteAccountRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/RelayAccountDeleteAccountRequestBody'
                    required:
                        - id
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/RelayAccountDeleteAccountResponseBody'
            schemes:
                - http
...
    /lpnode/lpnode_admin_panel/relayAccount/register:
        post:
            tags:
                - relayAccount
            summary: registerAccount relayAccount
            operationId: relayAccount#registerAccount
            parameters:
                - name: RegisterAccountRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/RelayAccountRegisterAccountRequestBody'
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/RelayAccountRegisterAccountResponseBody'
            schemes:
                - http
...
definitions:
...
    AuthenticationLimiterSetAuthenticationLimiterRequestBody:
        title: AuthenticationLimiterSetAuthenticationLimiterRequestBody
        type: object
        properties:
            authenticationLimiter:
                type: string
                example: Non nihil.
        example:
            authenticationLimiter: Neque distinctio dolor ut.
    AuthenticationLimiterSetAuthenticationLimiterResponseBody:
        title: AuthenticationLimiterSetAuthenticationLimiterResponseBody
        type: object
        properties:
            code:
                type: integer
                example: 2829824522465749384
                format: int64
            message:
                type: string
                example: Officiis consequatur temporibus.
            result:
                type: string
                description: result
                example: Omnis aspernatur.
        example:
            code: 8514865293155609618
            message: Modi et aut voluptatem eos illum.
            result: Explicabo rerum expedita odio.
...
    RelayAccountDeleteAccountRequestBody:
        title: RelayAccountDeleteAccountRequestBody
        type: object
        properties:
            id:
                type: string
                example: Dolores molestias et eos non ducimus odio.
        example:
            id: Ut repudiandae consequatur aut.
        required:
            - id
    RelayAccountDeleteAccountResponseBody:
        title: RelayAccountDeleteAccountResponseBody
        type: object
        properties:
            code:
                type: integer
                example: 697010819011822785
                format: int64
            message:
                type: string
                example: Consequatur nulla.
            result:
                type: string
                example: Quia occaecati.
        example:
            code: 1306682797554696621
            message: Consequatur consequatur similique molestias quo libero.
            result: Et magnam molestiae excepturi est aspernatur.
...
    RelayAccountRegisterAccountRequestBody:
        title: RelayAccountRegisterAccountRequestBody
        type: object
        properties:
            profile:
                type: string
                example: Eos ut aut.
        example:
            profile: Consequatur et sit voluptas.
    RelayAccountRegisterAccountResponseBody:
        title: RelayAccountRegisterAccountResponseBody
        type: object
        properties:
            code:
                type: integer
                example: 5312746738329648599
                format: int64
            message:
                type: string
                example: Odit eligendi quod.
            result:
                $ref: '#/definitions/relayAccountItemResponseBody'
        example:
            code: 2353024224465926760
            message: Rerum qui at odio unde deserunt.
            result:
                id: Omnis labore architecto ex eius ut.
                lpIdFake: Adipisci nesciunt provident quo.
                lpNodeApiKey: Fuga consequuntur fugiat id beatae ullam.
                name: Praesentium commodi error in.
                profile: Neque animi odio voluptatum atque qui.
                relayApiKey: Voluptatem ut quo maxime minus quis.
...
    relayAccountItemResponseBody:
        title: relayAccountItemResponseBody
        type: object
        properties:
            id:
                type: string
                example: Rerum accusantium.
            lpIdFake:
                type: string
                example: Dolorum provident asperiores sed asperiores at et.
            lpNodeApiKey:
                type: string
                example: Velit rerum quia reprehenderit quam.
            name:
                type: string
                example: Amet fuga architecto.
            profile:
                type: string
                example: Necessitatibus laudantium magni.
            relayApiKey:
                type: string
                example: Atque corporis.
        example:
            id: Doloremque inventore necessitatibus id consectetur minus.
            lpIdFake: In dolorum vero.
            lpNodeApiKey: Tempore sed eveniet.
            name: Molestiae eum eos blanditiis sunt sit nihil.
            profile: Incidunt mollitia quia nostrum accusamus fugiat.
            relayApiKey: Quis ut dolorem nobis voluptatem.
...